timesheet creation, timesheet review, timelog review, add configuration and environment variables for db, cloud, and payment gateway

+++ src/main/java/com/grownited/controller/Admin/AdminTaskController.java
@@ -51,7 +51,7 @@ public class AdminTaskController {
 	}
 	
 	@GetMapping("/deleteTask")
-	public String deleteTask(Integer taskId) {
+	public String deleteTask(Long taskId) {
 		
 		taskRepository.deleteById(taskId);
